desktopapp/requirements.txt
pycryptodome
Pillow
tk
qrcode
